+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Revision history for Z-Data
 
+## 0.1.6.1  -- 2020-10-09
+
+* Remove `Str` newtype.
+* Make `CBytes` a newtype.
+* Add JSON instances for `ExitCode`, Add Unaligned instances for `Ptr a`.
+* Use type alias instead of newtypes for `Locale`, `Category` in `Z.Data.Text`. 
+
 ## 0.1.6.0  -- 2020-10-09
 
 * Rename `ToText` to `ShowT`, `toText` to `showT`, add FFI types instances, remove `Str` newtype.